requirements.txt
Django==2.1
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
